fix(checker): TS1539/TS1273 grammar checks escape real-syntax-error suppression
Two independent bugs, both surfaced by the conformance snapshot refresh
(#16282) as "new failures" on bigintPropertyName.ts and jsdocTemplateTag7.ts,
though both predate this session's base and were simply uncaught before now.

**TS1539 (bigint literal property name).** tsc's checkGrammarProperty-family
diagnostics are suppressed program-wide once any file in the compilation has
a real syntax error (hasParseDiagnostics). tsz already models this via
`keep_checker_diagnostic_when_program_has_real_syntax_errors` /
`is_checker_routed_ts1xxx_grammar` for TS1064/TS1315/etc, but TS1539 was
never added to that list, so it survived program-wide suppression that the
real oracle (typescript-go 7.0.2) applies. Verified directly against the
pinned oracle: `bigintPropertyName.ts` alone reports many TS1539s; add a
sibling file with an unrelated genuine syntax error (`g.2n;`) and every
TS1539 disappears program-wide, oracle-side. Fix: add 1539 to
`is_checker_routed_ts1xxx_grammar` (crates/tsz-cli/src/driver/check_utils.rs).

**TS2304 for a JSDoc template parameter shadowed by an invalid modifier.**
`@template private T` correctly draws TS1273 (`'private' modifier cannot
appear on a type parameter`), but tsc still registers `T` as the function's
type parameter — it only rejects the modifier, not the parameter. tsz's two
name-extraction routines (`jsdoc_template_constraints` in
parsing_template_constraints.rs, `jsdoc_template_type_params` in
params_type_strings.rs) already skipped `const`/`in`/`out` modifier keywords
when scanning for the real name, but not the wider NEVER_VALID_MODIFIERS set
(private/public/protected/static/override/abstract/readonly/async/declare/
default/export) that `check_jsdoc_template_modifiers` uses for the TS1273
diagnostic itself. So `private` was read as the type parameter's own name,
`T` was never registered, and every use of `T` in `@param`/`@returns` drew a
spurious "Cannot find name 'T'". Verified against the oracle for `private`
and `static`: both fire only the modifier diagnostic, `T` still resolves.
Fix: hoist the modifier list to a shared
`NEVER_VALID_JSDOC_TEMPLATE_MODIFIERS` const and skip it in both
name-extraction sites, matching the existing `const`/`in`/`out` skip.

`parsing.rs` sat at 1998/2000 lines before this change, so the `@template`
constraint-extraction functions (`jsdoc_template_constraints`,
`jsdoc_template_constraints_before_typedef_host`) move verbatim (plus the
one-line modifier-skip extension) into a new
`jsdoc/parsing_template_constraints.rs`, a pure code move for the rest of
the diff.

## Goal
Goal: hold

## Verification
- `cargo test -p tsz-checker --lib jsdoc` (nextest) - 532/532, unchanged.
- `cargo test -p tsz-cli --lib check_utils` (nextest) - 106/106, unchanged.
- Targeted conformance, before/after this fix:
  - `compiler/bigintPropertyName.ts`: FAIL (extra TS1539 at 9 locations) -> PASS.
  - `conformance/jsdoc/jsdocTemplateTag7.ts`: FAIL (extra TS2304) -> PASS.
- Both failures verified directly against the pinned `typescript@7.0.2`
  oracle (isolated repros with/without a sibling syntax error; `private`
  and `static` JSDoc template modifiers) before coding, not just against the
  corpus baseline.
- `cargo fmt --all` clean; `cargo clippy --profile ci-lint -p tsz-checker -p
  tsz-cli --lib --tests -- -D warnings` clean; `python3
  scripts/arch/arch_guard.py` -> `Architecture guardrails passed`.
- Two-sided `scripts/conformance/compare-to-parent.sh origin/main` running;
  posting the number as a follow-up before queuing.
